+++ lib/blocs/theming/theming_state.dart
@@ -1,0 +1,43 @@
+import 'package:flex_color_scheme/flex_color_scheme.dart';
+import 'package:flutter/material.dart';
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+part 'theming_state.freezed.dart';
+
+final themes = {
+  'Material': FlexScheme.material,
+  'Material High Contrast': FlexScheme.materialHc,
+  'Deep Blue': FlexScheme.deepBlue,
+  'Flutter Dash': FlexScheme.flutterDash,
+  'Grey Law': FlexScheme.greyLaw,
+  'Indigo': FlexScheme.indigo,
+  'Mallard Green': FlexScheme.mallardGreen,
+  'Mandy Red': FlexScheme.mandyRed,
+  'Mango': FlexScheme.mango,
+  'Blue Whale': FlexScheme.blueWhale,
+  'Damask': FlexScheme.damask,
+  'Amber': FlexScheme.amber,
+  'Shark': FlexScheme.shark,
+  'Sakura': FlexScheme.sakura,
+  'San Juan Blue': FlexScheme.sanJuanBlue,
+  'Barossa': FlexScheme.barossa,
+  'Blumine Blue': FlexScheme.blumineBlue,
+  'Aqua Blue': FlexScheme.aquaBlue,
+  'Wasabi': FlexScheme.wasabi,
+  'Vesuvius Burn': FlexScheme.vesuviusBurn,
+  'Verdun Hemlock': FlexScheme.verdunHemlock,
+  'Outer Space': FlexScheme.outerSpace,
+  'Hippie Blue': FlexScheme.hippieBlue,
+  'Money': FlexScheme.money,
+};
+
+@freezed
+class ThemingState with _$ThemingState {
+  factory ThemingState({
+    ThemeData? darkTheme,
+    ThemeData? lightTheme,
+    String? darkThemeName,
+    String? lightThemeName,
+    ThemeMode? themeMode,
+  }) = _ThemingState;
+}