+++ test/pages/settings/about_page_test.dart
@@ -3,7 +3,6 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/pages/settings/advanced/about_page.dart';
-import 'package:lotti/themes/themes_service.dart';
 import 'package:mocktail/mocktail.dart';
 
 import '../../mocks/mocks.dart';
@@ -23,9 +22,7 @@ void main() {
       when(mockJournalDb.watchCountImportFlagEntries)
           .thenAnswer((_) => Stream<int>.fromIterable([0]));
 
-      getIt
-        ..registerSingleton<JournalDb>(mockJournalDb)
-        ..registerSingleton<ThemesService>(ThemesService());
+      getIt.registerSingleton<JournalDb>(mockJournalDb);
 
       when(
         () => mockJournalDb.watchTaskCount(any()),