+++ test/pages/settings/settings_page_test.dart
@@ -5,7 +5,6 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/pages/settings/settings_page.dart';
-import 'package:lotti/themes/themes_service.dart';
 import 'package:mocktail/mocktail.dart';
 
 import '../../mocks/mocks.dart';
@@ -22,9 +21,7 @@ void main() {
       when(mockJournalDb.watchJournalCount)
           .thenAnswer((_) => Stream<int>.fromIterable([n]));
 
-      getIt
-        ..registerSingleton<JournalDb>(mockJournalDb)
-        ..registerSingleton<ThemesService>(ThemesService());
+      getIt.registerSingleton<JournalDb>(mockJournalDb);
     });
     tearDown(getIt.reset);