+++ test/themes/utils_test.dart
@@ -3,9 +3,7 @@ import 'package:lotti/classes/tag_type_definitions.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/themes/colors.dart';
-import 'package:lotti/themes/themes_service.dart';
 import 'package:lotti/themes/utils.dart';
-import 'package:lotti/utils/consts.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
@@ -15,19 +13,7 @@ void main() {
     setUpAll(() {
       final db = JournalDb(inMemoryDatabase: true);
 
-      getIt
-        ..registerSingleton<JournalDb>(db)
-        ..registerSingleton(
-          ThemesService(),
-        );
-
-      db.insertFlagIfNotExists(
-        const ConfigFlag(
-          name: showBrightSchemeFlag,
-          description: 'Show Bright ☀️ scheme?',
-          status: false,
-        ),
-      );
+      getIt.registerSingleton<JournalDb>(db);
     });
     tearDownAll(() async {
       await getIt.reset();