+++ test/widgets/create/add_tag_actions_test.dart
@@ -4,7 +4,6 @@ import 'package:lotti/get_it.dart';
 import 'package:lotti/logic/persistence_logic.dart';
 import 'package:lotti/services/nav_service.dart';
 import 'package:lotti/services/time_service.dart';
-import 'package:lotti/themes/themes_service.dart';
 import 'package:lotti/widgets/create/add_tag_actions.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:mocktail/mocktail.dart';
@@ -25,7 +24,6 @@ void main() {
 
     setUp(() {
       getIt
-        ..registerSingleton<ThemesService>(ThemesService())
         ..registerSingleton<NavService>(mockNavService)
         ..registerSingleton<TimeService>(mockTimeService)
         ..registerSingleton<PersistenceLogic>(mockPersistenceLogic);