+++ test/widgets/journal/entry_details/measurement_summary_test.dart
@@ -5,7 +5,6 @@ import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/logic/health_import.dart';
 import 'package:lotti/services/entities_cache_service.dart';
-import 'package:lotti/themes/themes_service.dart';
 import 'package:lotti/widgets/journal/entry_details/measurement_summary.dart';
 import 'package:mocktail/mocktail.dart';
 
@@ -25,7 +24,6 @@ void main() {
       mockJournalDb = MockJournalDb();
 
       getIt
-        ..registerSingleton<ThemesService>(ThemesService())
         ..registerSingleton<JournalDb>(mockJournalDb)
         ..registerSingleton<EntitiesCacheService>(mockEntitiesCacheService)
         ..registerSingleton<HealthImport>(mockHealthImport);